Use float for number formatting in formatDecimal

+++ app/Traits/FormatNumbers.php
@@ -18,14 +18,14 @@ trait FormatNumbers
         $value = (string) $value;
         if (strpos($value, '.') !== false) {
             [$whole, $decimal] = explode('.', $value, 2);
-            $wholeFormatted = number_format((int) $whole);
+            $wholeFormatted = number_format((float) $whole);
             $decimal = rtrim($decimal, '0');
             if (strlen($decimal) < 2) {
                 $decimal = str_pad($decimal, 2, '0');
             }
             return "{$wholeFormatted}.{$decimal}";
         }
-        return number_format((int) $value, 2);
+        return number_format((float) $value, 2);
     }
     /**
      * Format a model attribute using formatDecimal.
